Adding a "Contribute" search scope
Also including hideScope, because once you turn on scope it is used in searches *but* the index doesn't know about a new scope for a day or so... so all the searches return zero results. Ideal technique is to turn on a scope, make it live but hidden, then go back and remove hideScope after 24-48 hours (after results are working with the Contribute scope).