Update in "Git source control section" (#128)
* Update in "Git source control section"

Added "on Pluralsight" for the Git course from Code School because it has now merged in the Pluralsight platform. 
Code School has closed its website & the link mentioned here redirects to a different Git course and also need to be a member in order to do that course, which does mention that a person can sign up for a free trial.
Not sure if you either want to keep the same link or change it to other Git courses offered on Pluralsight by Code School that are free to use.

* Update additional-resources.md

Added new link for the "Git course from Code School on Pluralsight" in the "Git Source Control Resources" section.
The name of the free course is called "Code School: Git Real".