Update for PowerApps languages
We are replacing the existing PowerApps entry that was unused with two new langauges: powerapps-dot and powerapps-comma.  There are two languages since we localize the formula language to be user friendly in regions that use "," (comma) as a decimal separator, consistent with what Excel does.  @DuncanmaMSFT has been helping us with setting this up and has reviewed this change.